backend/requirements.txt
pandas
numpy
scikit-learn
nltk
scikit-surprise
streamlit
pickle-mixin
